\name{GPmaps}
\docType{data}
\alias{GPmaps}
\alias{A}
\alias{D}
\alias{AA}
\alias{AD}
\alias{DA}
\alias{DD}
\alias{mouseweight}

\title{Dataset containing example GP maps}
     
\description{ Example GP maps with two loci including: The orthogonal GP maps \code{A}, \code{D}, \code{AA}, \code{AD}, \code{DA} and \code{DD}  used in 
decomposition of genetic variance (Zeng \emph{et al.} 2005). The GP map \code{mouseweight} for body weight studied by Cheverud \emph{et al.} (1995). }

\usage{GPmaps}

\format{Objects of class \code{\link{gpmap}}}
     
\references{
	
Cheverud JM & Routman EJ (1995) Epistasis and Its Controbution to Genetic Variance Components. 
Genetics 139:1455-1461 [\href{http://www.genetics.org/content/139/3/1455.abstract}{link}]
	
Zeng ZB, Wang T, Zou W. (2005). Modelling quantitative trait loci and interpretation of models. 
Genetics 169: 1711-1725. [\href{http://www.genetics.org/content/169/3/1711.full}{link}]
     }
     
\keyword{datasets}